langgraph-rag/requirements.txt
langgraph==0.2.39
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
fastapi==0.115.5
uvicorn==0.32.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1